+++ connector-integrationtest/src/test/java/eu/ecodex/connector/rest/admin/attachment/ConnectorListAttachmentsIT.java
@@ -20,6 +20,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.client.RestTestClient;
 
@@ -28,7 +30,8 @@ import org.springframework.test.web.servlet.client.RestTestClient;
     statements = "DELETE FROM connector_business_domains WHERE id > 0",
     executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS
 )
-public class ConnectorListAttachmentsIT extends AbstractIntegrationTest {
+class ConnectorListAttachmentsIT extends AbstractIntegrationTest {
+
     @Autowired
     private RestTestClient apiClient;
 
@@ -47,21 +50,25 @@ public class ConnectorListAttachmentsIT extends AbstractIntegrationTest {
         "classpath:sql/message.sql",
         "classpath:sql/message-as4-properties.sql",
         "classpath:sql/attachment.sql",
+        "classpath:sql/user.sql"
     })
     void should_list_attachments_for_connector_messages() {
         apiClient.get()
-                 .uri("/api/v1/admin/attachments")
-                 .exchange()
-                 .expectStatus().isOk()
-                 .expectBody(new ParameterizedTypeReference<ConnectorPageResult<ConnectorAttachmentDto>>() {
-                 })
-                 .value(result -> {
-                     assertThat(result).isNotNull();
-                     assert result != null;
-                     assertThat(result.content().size()).isEqualTo(14);
-                     assertThat(result.size()).isEqualTo(14);
-                     assertThat(result.totalElements()).isEqualTo(14);
-                     assertThat(result.totalPages()).isEqualTo(1);
-                 });
+            .uri("/api/v1/admin/attachments")
+            .header("Accept", MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + generateDefaultAdminToken())
+            .exchange()
+            .expectStatus().isOk()
+            .expectBody(
+                new ParameterizedTypeReference<ConnectorPageResult<ConnectorAttachmentDto>>() {
+                })
+            .value(result -> {
+                assertThat(result).isNotNull();
+                assert result != null;
+                assertThat(result.content().size()).isEqualTo(14);
+                assertThat(result.size()).isEqualTo(14);
+                assertThat(result.totalElements()).isEqualTo(14);
+                assertThat(result.totalPages()).isEqualTo(1);
+            });
     }
 }